refactor: Remove unused files and update package dependencies
